+++ server/src/main/java/com/shoppingsystem/shopping_system/service/CartItemService.java
@@ -4,6 +4,6 @@ import com.shoppingsystem.shopping_system.dto.CartItemDTO;
 import com.shoppingsystem.shopping_system.model.CartItem;
 
 public interface CartItemService {
-    CartItemDTO findById(int id);
+    CartItemDTO findById(Long id);
     CartItem save(CartItem cartItem);
 }
